requirements.txt
firebase-admin
aiocache~=0.12.3
py-cord~=2.6.0
pynacl
python-dotenv~=1.0.1
requests~=2.32.0
pillow~=11.0.0
pymupdf>=1.23.0
wavelink
#langchain
#langchain-community
#langchain-ollama
#langchain-chroma
#pypdf
#chromadb
#langdetect
sqlmodel==0.0.24
SQLAlchemy~=2.0.40
alembic
pydantic
aiohttp